pyproject.toml
[tool.poetry]
name = "tildejsongen"
version = "0.3.2"
description = "A simple Python script to generate a `tilde.json` file for Tilde-style servers"
authors = ["Andrew Williams <andy@tensixtyone.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
tildejsongen = 'tildejsongen:main'

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = ">= 5.4.1"
jinja2 = ">= 2.11"

[tool.poetry.group.test.dependencies]
ruff = "^0.11.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
lint.ignore = ["E501"]
